[update]レビュー詳細ページの改良
